requirements.txt
google-assistant-library==0.0.3
argparse==1.4.0
google-auth==1.0.1
RPi.GPIO==0.6.3